requirements.txt
streamlit 
groq 
python-pptx 
python-dotenv 
beautifulsoup4 
requests
mcp[cli]
langchain-mcp-adapters
langgraph
streamlit-reveal-slides